+++ tencentcloud/ess/v20201111/models.go
@@ -2565,6 +2565,9 @@ type CreateReleaseFlowRequestParams struct {
 	// 默认使用原流程的签署人列表,当解除协议的签署人与原流程的签署人不能相同时（例如原流程签署人离职了），需要指定本企业其他已实名员工来替换原流程中的原签署人，注意需要指明原签署人的编号(ReceiptId,通过DescribeFlowInfo接口获取)来代表需要替换哪一个签署人
 	// 解除协议的签署人数量不能多于原流程的签署人数量
 	ReleasedApprovers []*ReleasedApprover `json:"ReleasedApprovers,omitempty" name:"ReleasedApprovers"`
+
+	// 签署流程的签署截止时间。 值为unix时间戳,精确到秒,不传默认为当前时间七天后
+	Deadline *int64 `json:"Deadline,omitempty" name:"Deadline"`
 }
 
 type CreateReleaseFlowRequest struct {
@@ -2583,6 +2586,9 @@ type CreateReleaseFlowRequest struct {
 	// 默认使用原流程的签署人列表,当解除协议的签署人与原流程的签署人不能相同时（例如原流程签署人离职了），需要指定本企业其他已实名员工来替换原流程中的原签署人，注意需要指明原签署人的编号(ReceiptId,通过DescribeFlowInfo接口获取)来代表需要替换哪一个签署人
 	// 解除协议的签署人数量不能多于原流程的签署人数量
 	ReleasedApprovers []*ReleasedApprover `json:"ReleasedApprovers,omitempty" name:"ReleasedApprovers"`
+
+	// 签署流程的签署截止时间。 值为unix时间戳,精确到秒,不传默认为当前时间七天后
+	Deadline *int64 `json:"Deadline,omitempty" name:"Deadline"`
 }
 
 func (r *CreateReleaseFlowRequest) ToJsonString() string {
@@ -2601,6 +2607,7 @@ func (r *CreateReleaseFlowRequest) FromJsonString(s string) error {
 	delete(f, "NeedRelievedFlowId")
 	delete(f, "ReliveInfo")
 	delete(f, "ReleasedApprovers")
+	delete(f, "Deadline")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "CreateReleaseFlowRequest has unknown keys!", "")
 	}
@@ -5862,6 +5869,14 @@ type ReleasedApprover struct {
 	// ORGANIZATION-企业
 	// ENTERPRISESERVER-企业静默签
 	ApproverType *string `json:"ApproverType,omitempty" name:"ApproverType"`
+
+	// 签署控件类型，支持自定义企业签署方的签署控件为“印章”或“签名”
+	// - SIGN_SEAL-默认为印章控件类型
+	// - SIGN_SIGNATURE-手写签名控件类型
+	ApproverSignComponentType *string `json:"ApproverSignComponentType,omitempty" name:"ApproverSignComponentType"`
+
+	// 签署方自定义控件别名，最大长度20个字符
+	ApproverSignRole *string `json:"ApproverSignRole,omitempty" name:"ApproverSignRole"`
 }
 
 type RelieveInfo struct {